+++ ecscluster/main.tf
@@ -13,6 +13,8 @@ module "asg" {
   subnets              = "${var.subnets}"
   policies             = ["arn:aws:iam::aws:policy/service-role/AmazonEC2ContainerServiceforEC2Role"]
   user_data            = "${base64encode(data.template_file.instance_init.rendered)}"
+  volume_size = "${var.volume_size}"
+  volume_encryption = "${var.volume_encryption}"
   instance_schedule    = "${var.instance_schedule}"
   instance_patch_group = "${var.instance_patch_group}"
   instance_backup      = "${var.instance_backup}"