+++ rdsinstance/main.tf
@@ -19,6 +19,7 @@ resource "aws_db_instance" "default" {
   deletion_protection = true
   maintenance_window = "wed:04:00-wed:05:00"
   storage_encrypted = "${var.storage_encrypted}"
+  parameter_group_name = "${var.parameter_group_name}"
   db_subnet_group_name = "${aws_db_subnet_group.default.name}"
   vpc_security_group_ids = [
     "${var.security_groups}",